style and an extra check

+++ src/gui/Gui.cc
@@ -96,8 +96,8 @@ std::unique_ptr<ignition::gui::Application> createGui(
     defaultConfigFolder =
       ignition::common::joinPaths(defaultConfig, ".ignition",
         "gazebo", IGNITION_GAZEBO_MAJOR_VERSION_STR);
-    defaultConfig = ignition::common::joinPaths(defaultConfig, ".ignition",
-        "gazebo", IGNITION_GAZEBO_MAJOR_VERSION_STR, defaultGuiConfigName);
+    defaultConfig = ignition::common::joinPaths(defaultConfigFolder,
+        defaultGuiConfigName);
   }
   else
   {
@@ -271,15 +271,25 @@ std::unique_ptr<ignition::gui::Application> createGui(
     {
       if (!ignition::common::exists(defaultConfigFolder))
       {
-        if(!ignition::common::createDirectories(defaultConfigFolder))
+        if (!ignition::common::createDirectories(defaultConfigFolder))
         {
           ignerr << "Failed to create the default config folder ["
             << defaultConfigFolder << "]\n";
           return nullptr;
         }
       }
+
       auto installedConfig = ignition::common::joinPaths(
           IGNITION_GAZEBO_GUI_CONFIG_PATH, defaultGuiConfigName);
+      if (!ignition::common::exists(installedConfig))
+      {
+        ignerr << "Failed to copy installed config [" << installedConfig
+               << "] to default config [" << defaultConfig << "]."
+               << "(file " << installedConfig << " doesn't exist)"
+               << std::endl;
+        return nullptr;
+      }
+
       if (!ignition::common::copyFile(installedConfig, defaultConfig))
       {
         ignerr << "Failed to copy installed config [" << installedConfig
